+++ x-pack/plugins/apm/server/lib/apm_telemetry/make_apm_usage_collector.ts
@@ -5,13 +5,9 @@
  */
 
 import { CoreSetup } from 'src/core/server';
-import {
-  APM_TELEMETRY_DOC_ID,
-  createApmTelementry,
-} from './apm_telemetry';
-import { getSavedObjectsClient } from '../saved_objects/client';
+import { getSavedObjectsClient } from '../helpers/saved_objects_client';
+import { APM_TELEMETRY_DOC_ID, createApmTelementry } from './apm_telemetry';
 
-// TODO this type should be defined by the platform
 export interface CoreSetupWithUsageCollector extends CoreSetup {
   http: CoreSetup['http'] & {
     server: {
